secure_med_app/requirements.txt
# Core Flask
Flask==2.3.2

# Database ORM
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.3

# Database driver (for MySQL)
PyMySQL==1.1.0

# Migrations (optional but recommended)
Flask-Migrate==4.0.5

# Encryption
cryptography==41.0.7

# Environment variables
python-dotenv==1.0.1

# Testing & HTTP requests
requests==2.31.0
